Move normalize_deep test to chalk-engine
