+++ chalk-engine/src/normalize_deep.rs
@@ -105,3 +105,35 @@ where
         self.interner()
     }
 }
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use chalk_integration::interner::ChalkIr;
+    use chalk_integration::{arg, ty, ty_name};
+
+    const U0: UniverseIndex = UniverseIndex { counter: 0 };
+
+    #[test]
+    fn infer() {
+        let interner = &ChalkIr;
+        let mut table: InferenceTable<ChalkIr> = InferenceTable::new();
+        let environment0 = Environment::new(interner);
+        let a = table.new_variable(U0).to_ty(interner);
+        let b = table.new_variable(U0).to_ty(interner);
+        table
+            .unify(interner, &environment0, &a, &ty!(apply (item 0) (expr b)))
+            .unwrap();
+        assert_eq!(
+            DeepNormalizer::normalize_deep(&mut table, interner, &a),
+            ty!(apply (item 0) (expr b))
+        );
+        table
+            .unify(interner, &environment0, &b, &ty!(apply (item 1)))
+            .unwrap();
+        assert_eq!(
+            DeepNormalizer::normalize_deep(&mut table, interner, &a),
+            ty!(apply (item 0) (apply (item 1)))
+        );
+    }
+}
